+++ nzshm_model/psha_adapter/openquake/__init__.py
@@ -1,4 +1,14 @@
 """
+This package provides a psha_adapter implementation for the openquake PSHA.
+
+Classes:
+    OpenquakeConfig: a concrete psha_adapter for openquake hazard configuration
+
+Attributes:
+    DEFAULT_HAZARD_CONFIG: the base hazard configuration for NZHSM
+
 """
+from .hazard_config import OpenquakeConfig
+from .hazard_config_compat import DEFAULT_HAZARD_CONFIG
 from .logic_tree import NrmlDocument
-from .simple_nrml import OpenquakeSimplePshaAdapter
+from .simple_nrml import OpenquakeSimplePshaAdapter, gmcm_branch_from_element_text