+++ tests/test_logic_tree_correlations_new.py
@@ -14,7 +14,7 @@ def test_v2_source_tree_from_json_no_correlations():
     slt = SourceLogicTree.from_json(config)
 
     # without correlations there are 4 composite branches
-    assert len(list(slt.combined_branches)) == 4
+    assert len(list(slt.composite_branches)) == 4
 
     cor1 = Correlation(
         primary_branch=slt.branch_sets[0].branches[0],
@@ -27,7 +27,7 @@ def test_v2_source_tree_from_json_no_correlations():
     slt.correlations = LogicTreeCorrelations([cor1, cor2])
 
     # with correlations there are 2 composite branches
-    for branch in slt.combined_branches:
+    for branch in slt.composite_branches:
         print(branch)
         print('')
-    assert len(list(slt.combined_branches)) == 2
+    assert len(list(slt.composite_branches)) == 2